Network.Curl.Post

Description

Representing and marshalling formdata (as part of POST uploads/submissions.) If you are only looking to submit a sequence of name=value pairs, you are better off using the CurlPostFields constructor; much simpler.

Documentation

type Header = String

data HttpPost

Constructors

HttpPost

Fields

postName :: String

contentType :: Maybe String

content :: Content

extraHeaders :: [Header]

showName :: Maybe String

data Content

Constructors

ContentFile FilePath

ContentBuffer (Ptr CChar) Long

ContentString String

multiformString :: String -> String -> HttpPost

sizeof_httppost :: Int

marshallPosts :: [HttpPost] -> IO (Ptr HttpPost)

marshallPost :: HttpPost -> IO (Ptr HttpPost)

curl_slist_append :: Ptr Slist_ -> CString -> IO (Ptr Slist_)

curl_slist_free :: Ptr Slist_ -> IO ()

curl_formfree :: Ptr a -> IO ()